refactor: repair orchestration ownership boundaries
