+++ src/frame_compare/orchestration/execution.py
@@ -18,6 +18,18 @@ from frame_compare.analysis.metrics import ANALYZE_PROGRESS_TOTAL
 from frame_compare.config.schema import ConfigSchema
 from frame_compare.orchestration.analysis_policy import needs_analysis
 from frame_compare.orchestration.context import RunContext
+from frame_compare.orchestration.execution_types import (
+    ConfirmSlowpicsUploadPhaseOutput,
+    ExecutionPhasePlan,
+    ExecutionState,
+    MetadataPrefetch,
+    PhaseOutput,
+    PostReportCleanupPhaseOutput,
+    PrepState,
+    PublishPhaseOutput,
+    ReportPhaseOutput,
+)
+from frame_compare.orchestration.phase_output_application import apply_phase_output
 from frame_compare.orchestration.phase_post_render import (
     run_confirm_slowpics_upload_phase,
     run_metadata_phase,
@@ -36,20 +48,6 @@ from frame_compare.orchestration.phase_tasks import (
 )
 from frame_compare.orchestration.phases import Phase
 from frame_compare.orchestration.types import (
-    AlignPhaseOutput,
-    AnalyzePhaseOutput,
-    ConfirmSlowpicsUploadPhaseOutput,
-    ExecutionPhasePlan,
-    ExecutionState,
-    FramePlanPhaseOutput,
-    MetadataPhaseOutput,
-    MetadataPrefetch,
-    PhaseOutput,
-    PostReportCleanupPhaseOutput,
-    PrepState,
-    PublishPhaseOutput,
-    RenderPhaseOutput,
-    ReportPhaseOutput,
     RunDependencies,
     RunRequest,
     SlowpicsUploadConfirmationFn,
@@ -87,7 +85,7 @@ def _create_timed_phase(
                 output = await maybe_awaitable
             else:
                 output = maybe_awaitable
-            _apply_phase_output(ctx=ctx, state=state, output=output)
+            apply_phase_output(ctx=ctx, state=state, output=output)
         except Exception as exc:
             if warn_only:
                 warnings.append(f"{name}: {exc}")
@@ -104,59 +102,6 @@ def _create_timed_phase(
         progress_total=progress_total,
         warn_only=warn_only,
     )
-
-
-def _apply_phase_output(*, ctx: RunContext, state: ExecutionState, output: PhaseOutput) -> None:
-    match output:
-        case FramePlanPhaseOutput() as phase_output:
-            state.selected_frames[:] = phase_output.selected_frames
-            ctx.selection_breakdown = phase_output.selection_breakdown
-            ctx.selection_details_by_source_frame = phase_output.selection_details_by_source_frame
-            state.warnings.extend(phase_output.warnings)
-        case AnalyzePhaseOutput() as phase_output:
-            state.selected_frames[:] = phase_output.selected_frames
-            state.artifacts.metrics_cache_hit = phase_output.metrics_cache_hit
-            state.artifacts.metrics_cache_status = (
-                "hit" if phase_output.metrics_cache_hit else "miss"
-            )
-            ctx.selection_breakdown = phase_output.selection_breakdown
-            ctx.selection_details_by_source_frame = phase_output.selection_details_by_source_frame
-            ctx.analysis_metrics = phase_output.analysis_metrics
-        case AlignPhaseOutput() as phase_output:
-            ctx.reference = phase_output.reference
-            ctx.comparisons = phase_output.comparisons
-            state.selected_frames[:] = phase_output.selected_frames
-            state.warnings.extend(phase_output.warnings)
-            if phase_output.selection_breakdown is not None:
-                ctx.selection_breakdown = phase_output.selection_breakdown
-            if phase_output.selection_details_by_source_frame is not None:
-                ctx.selection_details_by_source_frame = (
-                    phase_output.selection_details_by_source_frame
-                )
-        case RenderPhaseOutput() as phase_output:
-            state.artifacts.render = phase_output.render
-            state.warnings.extend(phase_output.render.warnings)
-        case MetadataPhaseOutput() as phase_output:
-            state.artifacts.resolved_metadata = phase_output.resolved_metadata
-        case PublishPhaseOutput() as phase_output:
-            state.artifacts.slowpics_url = phase_output.slowpics_url
-            state.artifacts.uploaded_slowpics_file_paths = phase_output.uploaded_file_paths
-            state.artifacts.post_upload_actions = phase_output.post_upload_actions
-            state.warnings.extend(
-                action.warning
-                for action in phase_output.post_upload_actions
-                if action.warning is not None
-            )
-        case ReportPhaseOutput() as phase_output:
-            state.artifacts.report_path = phase_output.report_path
-            state.artifacts.report_succeeded = phase_output.report_succeeded
-        case ConfirmSlowpicsUploadPhaseOutput() as phase_output:
-            state.artifacts.slowpics_upload_confirmation_status = phase_output.status
-            state.warnings.extend(phase_output.warnings)
-        case PostReportCleanupPhaseOutput() as phase_output:
-            state.warnings.extend(phase_output.warnings)
-        case _:
-            raise TypeError(f"Unsupported phase output type: {output.__class__.__qualname__}")
 
 
 def build_phases_before_align(
